+++ crates/core/src/lib.rs
@@ -559,6 +559,7 @@ pub enum ArchiveError {
 
 pub trait ArchiveStore {
     type BlockIter<'a>: Iterator<Item = (BlockSlot, BlockBody)> + DoubleEndedIterator + 'a;
+    type SparseBlockIter: Iterator<Item = Result<(BlockSlot, Option<BlockBody>), ArchiveError>>;
 
     fn get_block_by_hash(&self, block_hash: &[u8]) -> Result<Option<BlockBody>, ArchiveError>;
 
@@ -574,6 +575,11 @@ pub trait ArchiveStore {
     fn get_tx(&self, tx_hash: &[u8]) -> Result<Option<EraCbor>, ArchiveError>;
 
     fn get_slot_for_tx(&self, tx_hash: &[u8]) -> Result<Option<BlockSlot>, ArchiveError>;
+
+    fn iter_blocks_with_address(
+        &self,
+        address: &[u8],
+    ) -> Result<Self::SparseBlockIter, ArchiveError>;
 
     fn get_range<'a>(
         &self,
@@ -605,6 +611,10 @@ pub enum MempoolError {
     #[cfg(feature = "phase2")]
     #[error("tx evaluation failed during phase-2: {0}")]
     Phase2Error(#[from] pallas::ledger::validate::phase2::error::Error),
+
+    #[cfg(feature = "phase2")]
+    #[error("phase-2 script yielded an error")]
+    Phase2ExplicitError,
 
     #[error("state error: {0}")]
     StateError(#[from] StateError),