+++ crates/minibf/src/routes/txs.rs
@@ -1,20 +1,22 @@
 use axum::{
-    Json,
     extract::{Path, State},
     http::StatusCode,
+    Json,
 };
 use blockfrost_openapi::models::{
     tx_content::TxContent, tx_content_cbor::TxContentCbor,
     tx_content_delegations_inner::TxContentDelegationsInner,
     tx_content_metadata_cbor_inner::TxContentMetadataCborInner,
-    tx_content_metadata_inner::TxContentMetadataInner, tx_content_utxo::TxContentUtxo,
+    tx_content_metadata_inner::TxContentMetadataInner, tx_content_mirs_inner::TxContentMirsInner,
+    tx_content_pool_certs_inner::TxContentPoolCertsInner,
+    tx_content_pool_retires_inner::TxContentPoolRetiresInner, tx_content_utxo::TxContentUtxo,
     tx_content_withdrawals_inner::TxContentWithdrawalsInner,
 };
 use dolos_core::{ArchiveStore as _, Domain};
 
 use crate::{
-    Facade,
     mapping::{IntoModel as _, TxModelBuilder},
+    Facade,
 };
 
 pub async fn by_hash<D: Domain>(
@@ -148,6 +150,59 @@ pub async fn by_hash_delegations<D: Domain>(
     let tx = TxModelBuilder::new(&raw, order)?
         .with_network(network)
         .with_chain(chain);
+
+    tx.into_response()
+}
+
+pub async fn by_hash_mirs<D: Domain>(
+    Path(tx_hash): Path<String>,
+    State(domain): State<Facade<D>>,
+) -> Result<Json<Vec<TxContentMirsInner>>, StatusCode> {
+    let hash = hex::decode(tx_hash).map_err(|_| StatusCode::BAD_REQUEST)?;
+
+    let (raw, order) = domain
+        .archive()
+        .get_block_with_tx(hash.as_slice())
+        .map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?
+        .ok_or(StatusCode::NOT_FOUND)?;
+
+    let network = domain.get_network_id()?;
+
+    let tx = TxModelBuilder::new(&raw, order)?.with_network(network);
+
+    tx.into_response()
+}
+
+pub async fn by_hash_pool_retires<D: Domain>(
+    Path(tx_hash): Path<String>,
+    State(domain): State<Facade<D>>,
+) -> Result<Json<Vec<TxContentPoolRetiresInner>>, StatusCode> {
+    let hash = hex::decode(tx_hash).map_err(|_| StatusCode::BAD_REQUEST)?;
+
+    let (raw, order) = domain
+        .archive()
+        .get_block_with_tx(hash.as_slice())
+        .map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?
+        .ok_or(StatusCode::NOT_FOUND)?;
+
+    let tx = TxModelBuilder::new(&raw, order)?;
+
+    tx.into_response()
+}
+
+pub async fn by_hash_pool_updates<D: Domain>(
+    Path(tx_hash): Path<String>,
+    State(domain): State<Facade<D>>,
+) -> Result<Json<Vec<TxContentPoolCertsInner>>, StatusCode> {
+    let hash = hex::decode(tx_hash).map_err(|_| StatusCode::BAD_REQUEST)?;
+
+    let (raw, order) = domain
+        .archive()
+        .get_block_with_tx(hash.as_slice())
+        .map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?
+        .ok_or(StatusCode::NOT_FOUND)?;
+
+    let tx = TxModelBuilder::new(&raw, order)?;
 
     tx.into_response()
 }