+++ crates/redb/src/archive/mod.rs
@@ -1,4 +1,5 @@
 use ::redb::{Database, MultimapTableHandle as _, Range, TableHandle as _};
+use redb::ReadTransaction;
 use std::path::Path;
 use tracing::{debug, info, warn};
 
@@ -164,11 +165,11 @@ impl ChainStore {
         }
     }
 
-    pub fn get_range<'a>(
+    pub fn get_range(
         &self,
         from: Option<BlockSlot>,
         to: Option<BlockSlot>,
-    ) -> Result<ChainIter<'a>, RedbArchiveError> {
+    ) -> Result<ChainRangeIter, RedbArchiveError> {
         let out = match self {
             ChainStore::SchemaV1(x) => x.get_range(from, to)?,
         };
@@ -209,6 +210,18 @@ impl ChainStore {
         }
     }
 
+    pub fn iter_blocks_with_address(
+        &self,
+        address: &[u8],
+    ) -> Result<ChainSparseIter, RedbArchiveError> {
+        match self {
+            ChainStore::SchemaV1(x) => {
+                // TODO: we need to filter the false positives
+                x.iter_possible_blocks_with_address(address)
+            }
+        }
+    }
+
     pub fn get_tx(&self, tx_hash: &[u8]) -> Result<Option<EraCbor>, RedbArchiveError> {
         match self {
             ChainStore::SchemaV1(x) => x.get_tx(tx_hash),
@@ -239,7 +252,8 @@ impl ChainStore {
 }
 
 impl dolos_core::ArchiveStore for ChainStore {
-    type BlockIter<'a> = ChainIter<'a>;
+    type BlockIter<'a> = ChainRangeIter;
+    type SparseBlockIter = ChainSparseIter;
 
     fn get_block_by_hash(&self, block_hash: &[u8]) -> Result<Option<BlockBody>, ArchiveError> {
         Ok(Self::get_block_by_hash(self, block_hash)?)
@@ -266,6 +280,13 @@ impl dolos_core::ArchiveStore for ChainStore {
 
     fn get_slot_for_tx(&self, tx_hash: &[u8]) -> Result<Option<BlockSlot>, ArchiveError> {
         Ok(Self::get_slot_for_tx(self, tx_hash)?)
+    }
+
+    fn iter_blocks_with_address(
+        &self,
+        address: &[u8],
+    ) -> Result<Self::SparseBlockIter, ArchiveError> {
+        Ok(Self::iter_blocks_with_address(self, address)?)
     }
 
     fn get_range<'a>(
@@ -295,9 +316,9 @@ impl From<v1::ChainStore> for ChainStore {
     }
 }
 
-pub struct ChainIter<'a>(Range<'a, BlockSlot, BlockBody>);
+pub struct ChainRangeIter(Range<'static, BlockSlot, BlockBody>);
 
-impl<'a> Iterator for ChainIter<'a> {
+impl Iterator for ChainRangeIter {
     type Item = (BlockSlot, BlockBody);
 
     fn next(&mut self) -> Option<Self::Item> {
@@ -308,12 +329,34 @@ impl<'a> Iterator for ChainIter<'a> {
     }
 }
 
-impl<'a> DoubleEndedIterator for ChainIter<'a> {
+impl DoubleEndedIterator for ChainRangeIter {
     fn next_back(&mut self) -> Option<Self::Item> {
         self.0
             .next_back()
             .map(|x| x.unwrap())
             .map(|(k, v)| (k.value(), v.value()))
+    }
+}
+
+pub struct ChainSparseIter(ReadTransaction, indexes::SlotKeyIterator);
+
+impl Iterator for ChainSparseIter {
+    type Item = Result<(BlockSlot, Option<BlockBody>), ArchiveError>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let next = self.1.next()?;
+
+        let Ok(slot) = next else {
+            return Some(Err(next.err().unwrap().into()));
+        };
+
+        let block = tables::BlocksTable::get_by_slot(&self.0, slot);
+
+        let Ok(block) = block else {
+            return Some(Err(block.err().unwrap().into()));
+        };
+
+        Some(Ok((slot, block)))
     }
 }